feat!: target ES6 JavaScript
BREAKING CHANGE: bump the target for compiled JavaScript from
ES5 to ES6, which should afford smaller code bundles and faster
runtimes thanks to newer JavaScript features provided by all
of the largest, modern browsers.

This is technically a breaking change but is expected to be of no
impact to most users.

Closes #133